+++ src/boot/proto/device-path.h
@@ -27,13 +27,14 @@ enum {
 
         HW_MEMMAP_DP                     = 0x03,
 
+        ACPI_DP                          = 0x01,
+
         MEDIA_HARDDRIVE_DP               = 0x01,
         MEDIA_VENDOR_DP                  = 0x03,
         MEDIA_FILEPATH_DP                = 0x04,
         MEDIA_PIWG_FW_FILE_DP            = 0x06,
         MEDIA_PIWG_FW_VOL_DP             = 0x07,
 
-        MSG_UART_DP                      = 0x0e,
         MSG_URI_DP                       = 24,
 };
 
@@ -47,6 +48,15 @@ typedef struct {
         EFI_DEVICE_PATH Header;
         EFI_GUID Guid;
 } _packed_ VENDOR_DEVICE_PATH;
+
+/* EISA PNP ID encoding: compressed 3-letter vendor + 16-bit product ID. */
+#define EISA_PNP_ID(Id) ((uint32_t) (((Id) << 16) | 0x41D0))
+
+typedef struct {
+        EFI_DEVICE_PATH Header;
+        uint32_t HID;
+        uint32_t UID;
+} _packed_ ACPI_HID_DEVICE_PATH;
 
 typedef struct {
         EFI_DEVICE_PATH Header;
